delete_one

+++ routes/ticket/deleteTicket.js
@@ -1,12 +1,18 @@
 import Ticket from "../../models/Ticket.js";
 
 const deleteTicket = async (req, res) => {
-    const { id } = req.body;  // Extract id from request body
+    const { id } = req.params;
     try {
-        await Ticket.findByIdAndDelete(id);
-        res.status(200).send({ success: true });
+        // Check if the id is valid
+        const existedTicket= await Ticket.findById(id);
+        if (!existedTicket) {
+            return res.status(404).json({ message: "Ticket not found!" });
+        }
+        // Delete the todo
+        await  Ticket.findByIdAndDelete(id);
+        return res.status(200).json({ message: "Ticket deleted successfully!" });
     } catch (error) {
-        res.status(500).send({ success: false, error });
+        return res.status(500).json({ message: error.message });
     }
     res.send('delete success');
 }
